Phase 1 front-end
Files for the initial front end, with limited functionality (i.e. the buttons do not actually link to anything and the database is not connected). Links between sites are needed, information needs to be added to the index page. All web files are named index to be able to work on them without the linkage. The file named index is the initial home page that can be accessed before login. Index_user_reg is the site for user registration, that may need additional fields.